#RandomForestInPython
Random forest algorithm is a one of the most popular and most powerful supervised Machine Learning algorithm in Machine Learning that is capable of performing both regression and classification tasks. this algorithm creates the forest with a number of decision trees.
In general, the more trees in the forest the more robust the prediction.